+++ vrp-core/src/construction/heuristics/evaluators.rs
@@ -5,7 +5,6 @@ mod evaluators_test;
 use std::sync::Arc;
 
 use crate::construction::heuristics::*;
-use crate::models::common::Cost;
 use crate::models::problem::{Job, Multi, Single};
 use crate::models::solution::{Activity, Leg, Place};
 use crate::models::{ConstraintViolation, GoalContext};
@@ -63,16 +62,16 @@ pub fn eval_job_insertion_in_route(
     }
 
     let route_costs = goal.estimate(&MoveContext::route(&insertion_ctx.solution, route_ctx, eval_ctx.job));
-    let best_known_cost = match &alternative {
-        InsertionResult::Success(success) => Some(success.cost),
-        _ => None,
-    };
 
-    if let Some(best_known_cost) = best_known_cost {
-        if let Either::Left(_) = eval_ctx.result_selector.select_cost(best_known_cost, route_costs) {
-            return alternative;
+    // analyze alternative and return it if it looks better based on routing cost comparison
+    let (route_costs, best_known_cost) = if let Some(success) = alternative.as_success() {
+        match eval_ctx.result_selector.select_cost(success.cost.clone(), route_costs) {
+            Either::Left(_) => return alternative,
+            Either::Right(route_costs) => (route_costs, Some(success.cost.clone())),
         }
-    }
+    } else {
+        (route_costs, None)
+    };
 
     eval_ctx.result_selector.select_insertion(
         insertion_ctx,
@@ -87,8 +86,8 @@ pub fn eval_job_constraint_in_route(
     eval_ctx: &EvaluationContext,
     route_ctx: &RouteContext,
     position: InsertionPosition,
-    route_costs: Cost,
-    best_known_cost: Option<Cost>,
+    route_costs: InsertionCost,
+    best_known_cost: Option<InsertionCost>,
 ) -> InsertionResult {
     match eval_ctx.job {
         Job::Single(single) => eval_single(eval_ctx, route_ctx, single, position, route_costs, best_known_cost),
@@ -102,8 +101,8 @@ pub(crate) fn eval_single_constraint_in_route(
     route_ctx: &RouteContext,
     single: &Arc<Single>,
     position: InsertionPosition,
-    route_costs: Cost,
-    best_known_cost: Option<Cost>,
+    route_costs: InsertionCost,
+    best_known_cost: Option<InsertionCost>,
 ) -> InsertionResult {
     if let Some(violation) =
         eval_ctx.goal.evaluate(&MoveContext::route(&insertion_ctx.solution, route_ctx, eval_ctx.job))
@@ -123,8 +122,8 @@ fn eval_single(
     route_ctx: &RouteContext,
     single: &Arc<Single>,
     position: InsertionPosition,
-    route_costs: Cost,
-    best_known_cost: Option<Cost>,
+    route_costs: InsertionCost,
+    best_known_cost: Option<InsertionCost>,
 ) -> InsertionResult {
     let insertion_idx = get_insertion_index(route_ctx, position);
     let mut activity = Activity::new_with_job(single.clone());
@@ -155,8 +154,8 @@ fn eval_multi(
     route_ctx: &RouteContext,
     multi: &Arc<Multi>,
     position: InsertionPosition,
-    route_costs: Cost,
-    best_known_cost: Option<Cost>,
+    route_costs: InsertionCost,
+    best_known_cost: Option<InsertionCost>,
 ) -> InsertionResult {
     let insertion_idx = get_insertion_index(route_ctx, position).unwrap_or(0);
     // 1. analyze permutations
@@ -183,7 +182,7 @@ fn eval_multi(
                         None,
                         service,
                         &mut activity,
-                        0.,
+                        Default::default(),
                         SingleContext::new(None, in1.next_index),
                     );
 
@@ -192,7 +191,7 @@ fn eval_multi(
                         let activity = shadow.insert(activity, srv_res.index);
                         let activities = concat_activities(in1.activities, (activity, srv_res.index));
                         return MultiContext::success(
-                            in1.cost.unwrap_or(route_costs) + srv_res.cost.unwrap(),
+                            in1.cost.unwrap_or_else(|| route_costs.clone()) + srv_res.cost.unwrap(),
                             activities,
                         );
                     }
@@ -223,11 +222,11 @@ fn analyze_insertion_in_route(
     insertion_idx: Option<usize>,
     single: &Single,
     target: &mut Activity,
-    route_costs: Cost,
+    route_costs: InsertionCost,
     init: SingleContext,
 ) -> SingleContext {
     let mut analyze_leg_insertion = |leg: Leg<'_>, init| {
-        analyze_insertion_in_route_leg(eval_ctx, route_ctx, leg, single, target, route_costs, init)
+        analyze_insertion_in_route_leg(eval_ctx, route_ctx, leg, single, target, route_costs.clone(), init)
     };
 
     match insertion_idx {
@@ -247,7 +246,10 @@ fn analyze_insertion_in_route(
             |lhs: &SingleContext, rhs: &SingleContext| {
                 eval_ctx
                     .result_selector
-                    .select_cost(lhs.cost.unwrap_or(f64::MAX), rhs.cost.unwrap_or(f64::MAX))
+                    .select_cost(
+                        lhs.cost.clone().unwrap_or_else(InsertionCost::max_value),
+                        rhs.cost.clone().unwrap_or_else(InsertionCost::max_value),
+                    )
                     .is_left()
             },
         ),
@@ -260,7 +262,7 @@ fn analyze_insertion_in_route_leg(
     leg: Leg,
     single: &Single,
     target: &mut Activity,
-    route_costs: Cost,
+    route_costs: InsertionCost,
     init: SingleContext,
 ) -> Result<SingleContext, SingleContext> {
     let (items, index) = leg;
@@ -286,11 +288,11 @@ fn analyze_insertion_in_route_leg(
                 return SingleContext::fail(violation, acc);
             }
 
-            let costs = eval_ctx.goal.estimate(&MoveContext::activity(route_ctx, &activity_ctx)) + route_costs;
-            let other_costs = acc.cost.unwrap_or(f64::MAX);
+            let costs = eval_ctx.goal.estimate(&MoveContext::activity(route_ctx, &activity_ctx)) + &route_costs;
+            let other_costs = acc.cost.clone().unwrap_or_else(InsertionCost::max_value);
 
             match eval_ctx.result_selector.select_cost(costs, other_costs) {
-                Either::Left(_) => SingleContext::success(activity_ctx.index, costs, target.place.clone()),
+                Either::Left(costs) => SingleContext::success(activity_ctx.index, costs, target.place.clone()),
                 Either::Right(_) => SingleContext::skip(acc),
             }
         })
@@ -313,14 +315,14 @@ struct SingleContext {
     /// Insertion index.
     pub index: usize,
     /// Best cost.
-    pub cost: Option<Cost>,
+    pub cost: Option<InsertionCost>,
     /// Activity place.
     pub place: Option<Place>,
 }
 
 impl SingleContext {
     /// Creates a new empty context with given cost.
-    fn new(cost: Option<Cost>, index: usize) -> Self {
+    fn new(cost: Option<InsertionCost>, index: usize) -> Self {
         Self { violation: None, index, cost, place: None }
     }
 
@@ -335,7 +337,7 @@ impl SingleContext {
     }
 
     #[allow(clippy::unnecessary_wraps)]
-    fn success(index: usize, cost: Cost, place: Place) -> Result<Self, Self> {
+    fn success(index: usize, cost: InsertionCost, place: Place) -> Result<Self, Self> {
         Ok(Self { violation: None, index, cost: Some(cost), place: Some(place) })
     }
 
@@ -358,21 +360,21 @@ struct MultiContext {
     /// Insertion index for next service.
     pub next_index: usize,
     /// Cost accumulator.
-    pub cost: Option<Cost>,
+    pub cost: Option<InsertionCost>,
     /// Activities with their indices.
     pub activities: Option<Vec<(Activity, usize)>>,
 }
 
 impl MultiContext {
     /// Creates new empty insertion context.
-    fn new(cost: Option<Cost>, index: usize) -> Self {
+    fn new(cost: Option<InsertionCost>, index: usize) -> Self {
         Self { violation: None, start_index: index, next_index: index, cost, activities: None }
     }
 
     /// Promotes insertion context by best price.
     fn promote(left: Self, right: Self) -> Result<Self, Self> {
         let index = left.start_index.max(right.start_index) + 1;
-        let best = match (left.cost, right.cost) {
+        let best = match (&left.cost, &right.cost) {
             (Some(left_cost), Some(right_cost)) => {
                 if left_cost < right_cost {
                     left
@@ -423,7 +425,7 @@ impl MultiContext {
 
     /// Creates successful insertion context.
     #[allow(clippy::unnecessary_wraps)]
-    fn success(cost: Cost, activities: Vec<(Activity, usize)>) -> Result<Self, Self> {
+    fn success(cost: InsertionCost, activities: Vec<(Activity, usize)>) -> Result<Self, Self> {
         Ok(Self {
             violation: None,
             start_index: activities.first().unwrap().1,