+++ vrp-core/src/construction/probing/repair_solution.rs
@@ -119,7 +119,7 @@ fn synchronize_jobs(
                         route_ctx,
                         single,
                         position,
-                        0.,
+                        Default::default(),
                         None,
                     );