+++ vrp-core/src/models/goal.rs
@@ -352,10 +352,17 @@ impl GoalContext {
     }
 
     /// Estimates insertion cost (penalty) of the refinement move.
-    pub fn estimate(&self, move_ctx: &MoveContext<'_>) -> Cost {
-        self.local_objectives
-            .iter()
-            .flat_map(|objectives| objectives.iter().map(|objective| objective.estimate(move_ctx)))
-            .fold(Cost::default(), |acc, other| acc + other)
+    pub fn estimate(&self, move_ctx: &MoveContext<'_>) -> InsertionCost {
+        self.local_objectives.iter().fold(InsertionCost::default(), |acc, objectives| {
+            objectives
+                .iter()
+                .map(|objective| objective.estimate(move_ctx))
+                .zip(acc.into_iter().chain(std::iter::repeat(Cost::default())))
+                .map(|(a, b)| {
+                    // NOTE we simply add cost values of two
+                    a + b
+                })
+                .collect()
+        })
     }
 }