+++ vrp-core/src/solver/search/local/exchange_swap_star.rs
@@ -3,7 +3,6 @@
 mod exchange_swap_star_test;
 
 use super::*;
-use crate::models::common::Cost;
 use crate::models::problem::Job;
 use crate::solver::search::create_environment_with_custom_quota;
 use crate::utils::Either;
@@ -146,12 +145,31 @@ fn create_route_pairs(insertion_ctx: &InsertionContext, route_pairs_threshold: u
 }
 
 /// Finds insertion cost of the existing job in the route.
-fn find_insertion_cost(search_ctx: &SearchContext, job: &Job, route_ctx: &RouteContext) -> Cost {
-    let original_costs = route_ctx.get_route_cost();
+fn find_insertion_cost(search_ctx: &SearchContext, job: &Job, route_ctx: &RouteContext) -> InsertionCost {
+    route_ctx
+        .route()
+        .tour
+        .index(job)
+        .and_then(|idx| {
+            assert_ne!(idx, 0);
 
-    let route_ctx = remove_job_with_copy(search_ctx, job, route_ctx);
+            let mut route_ctx = route_ctx.deep_copy();
+            route_ctx.route_mut().tour.remove(job);
+            search_ctx.0.problem.goal.accept_route_state(&mut route_ctx);
 
-    original_costs - route_ctx.get_route_cost()
+            // NOTE This is not the best approach for multi-jobs
+            let &(insertion_ctx, leg_selection, result_selector) = search_ctx;
+            eval_job_insertion_in_route(
+                insertion_ctx,
+                &EvaluationContext { goal: insertion_ctx.problem.goal.as_ref(), job, leg_selection, result_selector },
+                &route_ctx,
+                InsertionPosition::Concrete(idx - 1),
+                InsertionResult::make_failure(),
+            )
+            .into_success()
+            .map(|success| success.cost)
+        })
+        .unwrap_or_default()
 }
 
 /// Tries to find insertion cost for `insert_job` in place of `extract_job`.
@@ -200,9 +218,7 @@ fn find_top_results(
                 (InsertionResult::Success(_), InsertionResult::Failure(_)) => Ordering::Less,
                 (InsertionResult::Failure(_), InsertionResult::Success(_)) => Ordering::Greater,
                 (InsertionResult::Failure(_), InsertionResult::Failure(_)) => Ordering::Equal,
-                (InsertionResult::Success(left), InsertionResult::Success(right)) => {
-                    compare_floats(left.cost, right.cost)
-                }
+                (InsertionResult::Success(left), InsertionResult::Success(right)) => left.cost.cmp(&right.cost),
             });
 
             results.truncate(3);
@@ -239,7 +255,7 @@ fn choose_best_result(
         .enumerate()
         .fold((0, &failure), |(acc_idx, acc_result), (idx, result)| match (acc_result, result) {
             (InsertionResult::Success(acc_success), InsertionResult::Success(success)) => {
-                match search_ctx.2.select_cost(acc_success.cost, success.cost) {
+                match search_ctx.2.select_cost(acc_success.cost.clone(), success.cost.clone()) {
                     Either::Left(_) => (acc_idx, acc_result),
                     Either::Right(_) => (idx, result),
                 }
@@ -253,7 +269,7 @@ fn choose_best_result(
     } else {
         match result {
             InsertionResult::Success(success) => InsertionResult::Success(InsertionSuccess {
-                cost: success.cost,
+                cost: success.cost.clone(),
                 job: success.job.clone(),
                 activities: success.activities.iter().map(|(activity, idx)| (activity.deep_copy(), *idx)).collect(),
                 actor: success.actor.clone(),
@@ -302,16 +318,16 @@ fn try_exchange_jobs_in_routes(
         .iter()
         .flat_map(|outer_job| {
             let delta_outer_job_cost = find_insertion_cost(&search_ctx, outer_job, outer_route_ctx);
-            inner_jobs.iter().map(move |inner_job| (outer_job, inner_job, delta_outer_job_cost))
+            inner_jobs.iter().map(move |inner_job| (outer_job, inner_job, delta_outer_job_cost.clone()))
         })
         .collect::<Vec<_>>();
 
     // search phase
     let (outer_best, inner_best, _) = map_reduce(
         job_pairs.as_slice(),
-        |&(outer_job, inner_job, delta_outer_job_cost)| {
+        |(outer_job, inner_job, delta_outer_job_cost)| {
             if is_quota_reached() {
-                return (InsertionResult::make_failure(), InsertionResult::make_failure(), Cost::default());
+                return (InsertionResult::make_failure(), InsertionResult::make_failure(), InsertionCost::default());
             }
 
             let delta_inner_job_cost = find_insertion_cost(&search_ctx, inner_job, inner_route_ctx);
@@ -322,26 +338,26 @@ fn try_exchange_jobs_in_routes(
             let outer_result = choose_best_result(
                 &search_ctx,
                 outer_in_place_result,
-                outer_top_results.get(outer_job).unwrap().as_slice(),
+                outer_top_results.get(*outer_job).unwrap().as_slice(),
             );
 
             let inner_result = choose_best_result(
                 &search_ctx,
                 inner_in_place_result,
-                inner_top_results.get(inner_job).unwrap().as_slice(),
+                inner_top_results.get(*inner_job).unwrap().as_slice(),
             );
 
             let delta_cost = match (&outer_result, &inner_result) {
                 (InsertionResult::Success(outer_success), InsertionResult::Success(inner_success)) => {
-                    outer_success.cost + inner_success.cost - delta_outer_job_cost - delta_inner_job_cost
+                    &outer_success.cost + &inner_success.cost - delta_outer_job_cost - delta_inner_job_cost
                 }
-                _ => 0.,
+                _ => InsertionCost::default(),
             };
 
             (outer_result, inner_result, delta_cost)
         },
-        || (InsertionResult::make_failure(), InsertionResult::make_failure(), 0.),
-        |left, right| match compare_floats(left.2, right.2) {
+        || (InsertionResult::make_failure(), InsertionResult::make_failure(), InsertionCost::default()),
+        |left, right| match left.2.cmp(&right.2) {
             Ordering::Less => left,
             _ => right,
         },