+++ vrp-core/src/solver/search/recreate/recreate_with_blinks.rs
@@ -140,7 +140,7 @@ impl ResultSelector for BlinkResultSelector {
         }
     }
 
-    fn select_cost(&self, left: f64, right: f64) -> Either<f64, f64> {
+    fn select_cost(&self, left: InsertionCost, right: InsertionCost) -> Either<InsertionCost, InsertionCost> {
         let is_blink = self.random.is_hit(self.ratio);
 
         if is_blink || left < right {